BAU: rename check-email-fraud-block-service in line with conventions
Make casing consistent with other files
